change headliner article structure
in this case, the article is the header, image, and body text.
the header wasn't included initially.

additionally, the body text itself isn't another article.

nice rule of thumb: it gets the article tag if it would make sense to
include just the contents of that tag in my RSS reader. (obviously,
we're talking semantics, not syntax.) in my RSS reader, this particular
article would have its title, image, and body text, so that's what goes
in the article tag :)